Add second core file to test non async path
Non async path has been quite a bit quicker in terms of development time
but I don't want to give up on the challenge of actually building something
inside of async. What was I missing? For starters, it seemed that
when the score was initialized at -1 x, when building the "printable"
map, it errored out, but the async version never told me that. I suppose
I never looked in the right place. I'll start there, and move forward
tomorrow. Potentially finish without async as well