Finished part2 of day13... Without async
I'm not sure how I feel about ditching async.
At the end of the day, it's more functional and less reliant on "mutable" state

(It's not really mutable, it's just harder to introspect as things travel
inside of each of the channels)

Also, it's potentially possible that we can simply follow the ball and be fine.
Doing a dry run to make sure our paddle is always underneath it works perfect.
I may try the async path just following the ball with the paddle.