Finish day17 Part2
Apparently, not many people did part2 on the computer. Brute force seems
to work relatively quickly, so that's what we do.

That is:
A. Grab some string that will repeat later (i.e. routine A)
B. Drop all instances of this from the start
C. Repeat A & B for routine B and routine C

Also did some refactoring and removed some extraneous unneeded code off of
part2. Part 1 with the moving and the turning of the bot has some complexities
that might be unnecessary. Ehg. Thar be dragons.