Start on day 18 part1
Attempting to do BFS to search from some point to another point
Although.... because of how that works, each rotation equaling a layer
and us forgetting the connections between each layer, we can't record
what doors are where and what keys are blocked by what. So, we need to either
record the links and then construct the paths later or do a complete DFS
and do things that way instead. Rip.